Move auth related code into its own file and updated constants
