+++ bctw-api/src/apis/user_api.ts
@@ -110,11 +110,12 @@ const getUser = async function (
     return res.status(500).send(error.message);
   }
   const results: QueryResultRow = getRowResults(result, fn_name, true);
-  const signup = await critterbase.post(
-    'signup', 
-    {keycloak_uuid: results.keycloak_guid, system_user_id: String(results.id), system_name: 'CRITTERBASE'}
-  );
-  return res.send({...results, critterbase_user_id: signup.data.user_id});
+  const signup = await critterbase.post('signup', {
+    keycloak_uuid: results.keycloak_guid,
+    user_identifier: String(results.idir),
+    system_name: 'CRITTERBASE',
+  });
+  return res.send({ ...results, critterbase_user_id: signup.data.user_id });
 };
 
 /**
